+++ src/Dataflow/Serialization/Network/NetworkDescriptionSerialization.cc
@@ -28,6 +28,7 @@
 
 #include <Dataflow/Serialization/Network/NetworkDescriptionSerialization.h>
 #include <Dataflow/Serialization/Network/XMLSerializer.h>
+#include <boost/filesystem/operations.hpp>
 
 using namespace SCIRun::Dataflow::Networks;
 
@@ -41,4 +42,41 @@ void ToolkitFile::load(std::istream& istr)
 void ToolkitFile::save(std::ostream& ostr) const
 {
   XMLSerializer::save_xml(*this, ostr, "toolkit");
+}
+
+namespace
+{
+  boost::filesystem::path diffPath(const boost::filesystem::path& basePath, const boost::filesystem::path& newPath)
+  {
+    namespace fs = boost::filesystem;
+
+    fs::path diffpath;
+
+    auto tmpPath = newPath;
+    while (tmpPath != basePath)
+    {
+      diffpath = tmpPath.stem() / diffpath;
+      tmpPath = tmpPath.parent_path();
+    }
+
+    auto filename = diffpath.leaf().string() + newPath.extension().string();
+    diffpath.remove_leaf() /= filename;
+    return diffpath;
+  }
+}
+
+ToolkitFile SCIRun::Dataflow::Networks::makeToolkitFromDirectory(const boost::filesystem::path& toolkitPath)
+{
+  ToolkitFile toolkit;
+
+  for (const auto& p : boost::filesystem::recursive_directory_iterator(toolkitPath))
+  {
+    if (p.path().extension() == ".srn5")
+    {
+      auto path = diffPath(toolkitPath, p.path()).string();
+      std::replace(path.begin(), path.end(), '\\', '/');
+      toolkit.networks[path] = *XMLSerializer::load_xml<NetworkFile>(p.path().string());
+    }
+  }
+  return toolkit;
 }
