+++ src/Dataflow/Serialization/Network/NetworkDescriptionSerialization.h
@@ -44,7 +44,7 @@ namespace SCIRun {
 namespace Dataflow {
 namespace Networks {
 
-  typedef std::vector<ConnectionDescriptionXML> ConnectionsXML;
+  using ConnectionsXML = std::vector<ConnectionDescriptionXML>;
 
   struct SCISHARE ModuleWithState
   {
@@ -65,7 +65,8 @@ namespace Networks {
   {
     std::string noteHTML, noteText;
     int position, fontSize;
-    NoteXML(const std::string& html = "", int p = 0, const std::string& text = "", int f = 12) :
+
+    explicit NoteXML(const std::string& html = "", int p = 0, const std::string& text = "", int f = 12) :
       noteHTML(html), noteText(text), position(p), fontSize(f) {}
   private:
     friend class boost::serialization::access;
@@ -172,6 +173,8 @@ namespace Networks {
       ar & BOOST_SERIALIZATION_NVP(networks);
     }
   };
+
+  SCISHARE ToolkitFile makeToolkitFromDirectory(const boost::filesystem::path& toolkitPath);
 
   template <class Value>
   std::map<std::string, Value> remapIdBasedContainer(const std::map<std::string, Value>& keyedByOriginalId, const std::map<std::string, std::string>& idMapping)
